refactor platedhole into it's own file
